[TASK] Make tabs context aware
We are dropping the Bootstrap CSS definitions for
all navs to have more control about the styling of
the nav elements. From now on we are shipping
custom css adapted to our needs. The new nav
element styling is context aware and adapts to
light and dark color schemes.

Resolves: #104865
Releases: main
Change-Id: I8632b33c48162cbb407dc2b2a99eaab00518c635
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85936
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>