[TASK] Make modal context aware
We are dropping the Bootstrap CSS definitions for
all modals to have more control about the styling of
the modal elements. From now on we are shipping
custom css adapted to our needs. The new modal
element styling is context aware and adapts to
light and dark color schemes.

Resolves: #104911
Releases: main
Change-Id: I98d619c407708f6d5c969cb98fc304786e4c0c44
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86015
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Benjamin Franzke <ben@bnf.dev>