[TASK] Make list-group context aware
We are dropping the Bootstrap CSS definitions for
all list-group to have more control about the styling of
the list-group elements. From now on we are shipping
custom css adapted to our needs. The new list-group
element styling is context aware and adapts to
light and dark color schemes.

Resolves: #104927
Releases: main
Change-Id: I9212f0e668a00591f1351b1a56daec99eadc420b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86041
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>