Revert "[FEATURE] Auto-discover record preview templates"
This reverts commit 3963de080e83a4d31531726469282d9a6941627f.

Reason for revert: Template auto discovery is a feature that needs
to designed with a full concept for the entire system in mind and
not for one single core detail. It should be part of a lower
"near fluid" level. This detail feature can be easily done as an
extension if people love it so much that they want it now.

Change-Id: I2d1e7eda2c1edfcce910b7ed2f5835508d4a26e4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86057
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>