[BUGFIX] Use correct element reference in mass action file upload
If a file with a filename is about to get uploaded that already exist
in the file list, a modal appears that asks the user what to do with
the conflicting file.

The event handler bound to the mass action select element now references
the correct event target.

Resolves: #105090
Releases: main
Change-Id: Ia46cff21300446b06285d0a859245baf0dc7f095
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86265
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>