[BUGFIX] Auto-select tree nodes in module deeplinks
Select pages and folders in the respective trees when opening module
routes via shared deeplinks, topbar bookmarks or via the opendocs menu.
This also implicitly retains the current tree selection when reloading
the TYPO3 backend.

The module state handling is slightly changed to better isolate
responsibility for the tree selection states:
 * modules are allowed to update the `identifier` property
   only (reflecting the current page id or folder identifier)
 * `treeIdentifier` property is the sole domain of the
   tree components (only the tree can know, which node –
   in case nodes are mounted multiple times – is to be selected)

Various tree components are updated to prepare nodes not only during
initial load but also during runtime (in order for selection to work in
any case) and have a post-initialization check to load nodes recursively
if the currently active node (per module state storage) is not visible.

Out-of-scope for this patch:
Tree nodes can not be auto-opened if a FormEngine view is loaded
(or any other non-module backend controller) as there is no module
and tree state attached to FormEngine, therefore shared Fomengine
deeplinks currently lack module context.

Releases: main, 13.4
Resolves: #104457
Resolves: #105120
Change-Id: I3d7d7876403259d4e1e234d43f4bff37b3c85d20
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86338
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benni Mack <benni@typo3.org>