[TASK] Enhance color switcher to a smart toggle w/ advanced options
Implement a color-scheme toggle button, that toggles between light
and dark color scheme, but takes "auto" into account by toggling
based on the auto-detected (dark/light) mode, when current mode is auto,
and toggles to "auto" when the to-state matched the auto-detected state
as this transition basically means user wants the color scheme to match
the current system settings (and is very likely fine if the scheme
is updated automatically, when the user changes the light/dark setting
on system level).

Additionally an expander is added, in order for all options
(including auto-detect) to be selectable within the user toolbar menu.

We have noticed that technical and subjective labels like "OS default"
or "User agent preference" are not perceived very well.
Therefore we removed "OS Default" from the toolbar in #104942.

Now this rather unintuitive option was gone from the user toolbar –
which was good for users that just wanted to toggle – but bad for power
users that like to enable "auto", but found it hard to discover the
auto option withing the user settings tab "Backend appearance".

Therefore the switcher now has an embedded advanced mode, where the
the "auto" item is available again and the label is combined of
"Auto detect" and the (auto detected) mode that it will be changed to.

The advantage is that there is less surprise when clicking auto detect,
when switching from "Dark mode" to "Auto detect (Dark mode)" or from
"Light mode" to "Auto detect (Light mode)".

Releases: main
Resolves: #105134
Change-Id: I748389bedb02d3cdfeefe0e0efcd0b30712c1b5b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86351
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>