[TASK] Use TCA schema for fetching relevant fields
Both ElementInformationController as well as RemoteServer previously
booted up FormEngine to fetch relevant fields for being displayed in
corresponding views.

This is bad in multiple terms: performance, misuse of FormEngine,
dependency to PSR-7, etc.

Therefore, a new VisibleSchemaFieldsCollector is introduced, which
uses the TCA schema to provide the fields to be displayed, depending
on the record type, the language context and the current backend user.

Resolves: #105185
Releases: main
Change-Id: I1e39611db467105ddaf2427dc47c7bd94ef7a915
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86405
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>