[TASK] Deprecate TCA sub types
The sub types feature of TCA was used for
registration of plugins as "list_type". Since
it's encouraged to register plugins as CType
since a couple of versions already and the
deprecation of the "list_type" functionality
with #105076 are the sub types now deprecated
as well.

This special functionality just lead to
headaches and should be replaced by using
dedicated record types, making configuration
much cleaner and more comprehensible.

Resolves: #105213
Related: #105076
Releases: main
Change-Id: If9ce3f5de7ef69aaf16cf987ebcb62ea50f9a6a6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86425
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>