[BUGFIX] Fix inline SVG login logo location for subdir installation
The SVG login logo contents will now be read from the original file pointer instead
of the resolved absolute public URL in order to be accessible for local
processing in a sub-directory installation.

Reason: The resolved absolute public folder of a file path includes the
sub-directory prefix. This absolute(!) URL must not be used as a
relative(!) path within the sub-directory itself. Otherwise, the sub-directory is tried to be retrieved within the sub-directory itself once more.

Example of incorrect absolute URL usage as a relative path:

  server docroot:
  /var/www/html/

  installation path:
  /var/www/html/my_dir/

  resolved URL:
  /my_dir/typo3/sysext/core/…/typo3_variable.svg

  incorrectly assumed local SVG path:
  /var/www/html/my_dir/my_dir/typo3/sysext/core/…/typo3_variable.svg
                ^^^^^^^^^^^^^
                |  invalid  |
                -------------

Resolves: #105205
Releases: main
Change-Id: I33f00d172ec0d375b90f79e163a0ba78ec65696f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86433
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>