[BUGFIX] Fix virtual embedded install tool url in subdir installation
The embedded install tool is dispatched from a virtual module URL –
{entryPoint}/module/tools/… – to {instancePath}typo3/index.php.

The backend module router takes care to virtually replace
{instancePath}typo3/index.php by {entryPoint}/module/tools/
in order for browser-reload to work.

When this virtual is created, the instance path needs to be
taken into accounts instead of implying TYPO3 path is `/`,
which is not the case in a sub directory installation.

Releases: main
Resolves: #105206
Change-Id: Ic257a16242deaa47ecd2f26bee91a47ed96bf49b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86434
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>