[BUGFIX] Suspend FormEngine validation on record type change
If a user attempts to change a record type, the backend user is asked to
save the form as it needs to get re-rendered. This will trigger the
FormEngine validation. If a user decided to leave out a field on
purpose, because it's not relevant in the new record type, FormEngine
validation will fail and stop the request, asking the user to fulfill
the field's constraints.

Example:

1. User creates a new "Page" record of type "External URL" by accident
2. User wants to switch to "Shortcut"
3. User has to fill the field "URL", even if it's not required by
   "Shortcut"

FormEngine validation currently has no sustainable solution for this
scenario. One solution could be to check whether the field *will be*
required in the new record type nevertheless.

This issue may cause major pain to editors. To workaround this issue for
now, validation is completely suspended when a FormEngine update has
been requested.

Resolves: #104996
Resolves: #105093
Releases: main
Change-Id: I0bc1b70c7fde14e22980cf8dc133ac71e480a361
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86526
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>