[TASK] Instantiate JavaScriptItems with new()
Class TYPO3\CMS\Core\Page\JavaScriptItems is a DTO.
It is a contract between PHP and JS with some
functionality on the PHP side to gather and prepare
the data to transfer and to park state on PHP side.
As such, this class can not be extended by extensions
since that would break the contract.
The patch thus declares the class final and changes
instantiation to use new().

Resolves: #105246
Releases: main
Change-Id: I312926489f529208fa067c4b174556a6f7db9c52
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86527
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>