[TASK] De-duplicate type update confirmation
In 2021 the FormEngine reload mechanism was reworked to trigger the
confirmation via a custom web component. This is not necessary anymore,
as half a year later the `ReloadOnFieldChange` form behavior class was
introduced.

This commit removes the custom web component in favor of
`ReloadOnFieldChange`, de-duplicating the logic and removing the need to
load additional assets.

Resolves: #105249
Related: #93899
Related: #91787
Releases: main
Change-Id: Ie59b86576d81c0dc98d567e9f3035923a67d5b4e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86529
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>