[BUGFIX] Prevent collapse state issues in pagetree filter
The pagetree has been changed to store collapse state in
local storage with #10528 but failed to adapt the state
handling for the filter mode, which caused pages to be
rendered without children but bing shows as being expanded.
This means the currently collapsed page had to be clicked
twice to toggle the state twice and to show children.

Also remove any backend code that is a left-over from the
backend-side state handling (UC) and instead calculate
`expanded` (now renamed to `__expanded` to reflect that it
is a client-side calculated property, synchronized with local
storage).

Releases: main
Resolves: #105248
Related: #103043
Change-Id: I90810e573913079a0494eb2c3a59ce4d543422c4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86531
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>