[BUGFIX] Render constants in tags in TypoScript BE modules
FineDiffViewHelper is used in Page TSconfig and TypoScript
"active" modules. That view helper switched from using
FineDiff directly to the core facade DiffUtility with #104325,
and added strip_tags() to the to-diff values.
When constants are used within TypoScript that looks like a
HTML tag, those are now stripped away, ending up with empty
"Constant usage:" rendering in the modules.

The patch fixes the view helper and activates output escaping
to be on the safe side.

Resolves: #104726
Related: #104325
Releases: main
Change-Id: I86ca507c3050e0f56667a7e7f255a1c1f76d02cb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86605
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>