[BUGFIX] Make popovers context aware
We are dropping the Bootstrap CSS definitions for
all popovers to have more control about the styling of
the popover elements. From now on we are shipping
custom css adapted to our needs. The new popover
element styling is context aware and adapts to
light and dark color schemes.

Resolves: #105374
Releases: main, 13.4
Change-Id: Iabe18fdf1c1d7190aa5f76679a37ab43baecb7f3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86626
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>