[BUGFIX] Use correct dataset identifier for date ranges
Datetime fields can be configured to set a minimum date and a maximum
date. This apparently broke with the migration to flatpickr, which
dropped jQuery in the same run.

The data attributes are named `data-date-min-date` and
`data-date-max-date`, and therefore the dataset accessor must be
correctly camel-cased. Configured date ranges are transformed to valid
JavaScript `Date` objects and fed into flatpickr.

To ensure consistent date handling, the min and max range values
are now parsed into server localtime, to coincide with how
the date values are handled since #104309.

Also, a minor CSS bug is fixed that makes disallowed dates
distinguishable from dates in another months.

Resolves: #105400
Related: #91606
Releases: main, 13.4, 12.4
Change-Id: Id6f1561e8b4d3b048abc883f8397507972e2d6d7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86829
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>