[BUGFIX] Make `GridEditor.writeConfig` update content area as intended
This fixes `GridEditor.writeConfig` to update the content area as
intended.

Before the code used an `undefined` property:
`codemirror.editorView.viewState.state.doc.length`

Correct is:
`codemirror.editorView.state.doc.length`

As a separate change made "t3-editor" part of the backend, the function
`CodeMirrorElement.setContent` is used, instead of fixing the issue.
Thereby the code is de-duplicated.

As LitElement `Ref`-objects are in use, first there is a check for
instantiation of the `Ref.value` to see if the actual object is
ready-to-use.

This also implements proper type-hints:

```
const codemirror: CodeMirrorElement | undefined = this.codeMirrorRef.value;
```

Instead of:

```
const codemirror: any = this.codeMirrorRef.value!;
```

Also this fixes web-test-runner import-map to correctly resolve scripts
in `backend/Resources/Public/JavaScript/Contrib`.

Resolves: #105422
Related: #105423
Releases: main, 13.4
Change-Id: I86b44788991922ba50d7fd8570b273b267e974ad
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87480
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>