[TASK] Return string in FormEngine `integer` parsing
The method `FormEngineValidation.processValue()` is declared to return a
`string`, but that's not always the case: in case for `integer` fields,
a real `number` is returned.

Fixing this reveals another misbehavior: if a number field contains the
value "42" and the field is updated to contain the value "0042", its
value normalized to the string value `'42'` again. Due to the strict
type check for value updates, the value is not considered as changed,
and therefore the "human-readable" field was not updated with the
normalized (formatted) value.

This commit therefore changes two things:

* string values are now returned when parsing `integer` fields
* the user facing field is now always updated to contain a clean value

Resolves: #105365
Releases: main, 13.4
Change-Id: I0b905876d3be984d2cad54c88e84e653b0fe58fa
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86833
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>