[BUGFIX] Children of selectTree filter should be expandable
This patch changes the behaviour of TCA renderType selectTree
filter so that it is possible to see unfiltered children of
nodes for which the filter matches. Previously, all nodes
were hidden if the filter did not match them.

Now, the behaviour initially looks the same, but it is possible
to expand the lowest filtered node and see the unfiltered
children (as in the page tree).

Resolves: #105609
Releases: main, 13.4
Change-Id: I48e09d0f253bb4a108e718ea667895b454da15f3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87005
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Andreas Nedbal <andy@pixelde.su>