[BUGFIX] Prevent BackendUtility::getRecord() postgresql out of range
This is a pre-patch for #94562: It has been discovered that
on postgresql a BackendUtility::getRecord() lookup with
a uid larger than the integer range raises a DriverException.

However, DataHandler and softrefParsers will only expect
a null response from such a query with UID ranges out of
bounds to carry on further exemptions (which is why
throwing an exception is not helpful for this).

This patch adds a test verifying this failure and
adds an early return in case of out of bound lookups.

Resolves: #105658
Related: #94562
Releases: main, 13.4, 12.4
Change-Id: Id10a0b31303be3890579723cdce644a3181451ef
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87105
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>