[BUGFIX] Consider virtual tables in ElementInformationController
When calculating the field list to be shown in
the ElementInformationController, virtual tables
like `_FOLDER` or `_FILE` are now properly handled
as they do not have a schema (TCA definition).

Additionally, this fixes the fact that the
expand / collapse state is now also saved
for folders. This previously was not recognized
since the element info did not work for folders
at all.

Resolves: #105676
Related: #105185
Releases: main, 13.4
Change-Id: Ia73512e4a499f9822ba3725f51d381332a7f4053
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87162
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>