[BUGFIX] Ignore invalid tree state in page tree drag actions
The expansion state of tree nodes is persisted in localStorage.
When all children of a node are removed (either locally or by
someone else) the state is still present and the nodes are
"enhanced" with the broken state later. This prevents dragging
elements after the now childless nodes.

This can be fixed manually with either clearing the tree state
localStorage or using the "Collapse all tree items" option, which
sets new state.

This fix in code however checks if the condition that sets the
expanded state again is only ever called if the to-be-enhanced
node actually has children, effectively ignoring the broken
state.

Resolves: #105680
Releases: main, 13.4
Change-Id: Icff6aeb21323187b2582d9f6f3deb848369a3125
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87168
Tested-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>