[BUGFIX] Expose record information in edit view
Currently, record information is only accessible via the context
menu on the record icon in the footer, which is unintuitive and
offers poor UX. To improve usability, essential actions are now
provided as dedicated buttons in the document header.

This change also aligns with the planned removal of the context
menu from the form engine footer, which includes actions unsuitable
for this view.

Resolves: #105764
Releases: main, 13.4
Change-Id: Ifbbdc34fc68406aa3a1921267abcae974aade4e3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87359
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Andreas Kienast <akienast@scripting-base.de>