[TASK] Remove context menu from formengine footer icon
The context menu on the footer icon exposes actions, such
as copy and paste, that are not appropriate for this view.
Important actions like "Show History" and "Record Information"
have already been made directly accessible in the editing view
with #105764 and #105763. The remaining actions in the context
menu offer little value and only add confusion, so the menu has
been removed.

Additionally, the title for the icon has been removed, as it is
already hidden from screen readers and the record type is
displayed next to it.

Resolves: #105770
Releases: main, 13.4
Change-Id: Ifcef786f5247c51de3b41f82e58e3304220e3511
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87505
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>