[BUGFIX] Respect order of item groups in NewContentElementWizard
Since the change of registering wizard items groups with TCA in #102834
the order of groups can't be changed anymore in the wizard.

This is addressed now with a rather simple solution: When TCA groups are
added, the sorting is already ensured by EMU:addTcaSelectItemGroup. We
take this sorting and auto-add the positional information in
NewContentElementController by list-linking the groups by
"after:{theGroupBefore}". This way the sorting is well-defined.

How to deal with PageTs: Well, we can't have it both ways. As soon
as PageTs overrides at least one group order, this will overrule the
TCA sorting. Then, everything has to be defined via PageTs. This is
fine, as we strive to reduce usage of PageTs.

Resolves: #104855
Related: #102834
Releases: main, 13.4
Change-Id: I071a74fd03b9a5e9ab4d4e471ad6226e485657da
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87382
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>