[FEATURE] Notify backend user on failed MFA verification attempts
In terms of security, it is best practice to notify the user on
failed MFA verification attempts. TYPO3 currently only logs failed
MFA verification attempts to the sys_log, resulting in the users
not being informed about the failed MFA verification attempt.

This change adds a notification mechanism to inform TYPO3 backend
users via email when an MFA provider is configured and an
incorrect MFA verification code is entered during login. The
notification is only sent if the backend user has a valid email
address configured.

Resolves: #105783
Releases: main
Change-Id: I6db53d968bc644c102ffae443284db9bfeaa20b3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87391
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>