[BUGFIX] Improve usability for selectTree
(1) Change input type to "search" in selectTree filter

Change input type to "search" in selectTree filter. This way, it is
possible to X out the text. This is more consistent with the other
fields like this, such as the page tree filter.

(2) Make language labels available

Some language labels were already been used (as title attributes
for the buttons and as placeholder text in the input field) but
the texts were not defined. The language labels are added to
locallang_alt_doc.xlf. The buttons now show a short description
on hover.

Resolves: #105789
Resolves: #105790
Releases: main, 13.4
Change-Id: If6e8ec0910213d70f684ba4402dbaf32dd25c360
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87465
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>