[BUGFIX] Register consecutively created drag uploaders
The jQuery removal in #102415 removed the MutationObserver, breaking the
registration of consecutively created drag uploaders, e.g. in IRRE
items.

The MutationObserver is added again to solve this issue, but reveals
that the code needs a complete overhaul as needing a MutationObserver
for this use-case is too much.

Resolves: #105755
Related: #102415
Releases: main, 13.4
Change-Id: I5dc2766d7a3a960903493b66332e874ebf9d3650
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87440
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>