[BUGFIX] Adjust display of form wizards button of form field 'group'
The field type 'group' shows the selected elements below the field. If
an element has a longer title and the viewport was no longer sufficient
for a complete display, this resulted in the action buttons no longer
being visible and the editing area being expanded. This patch fixes
this behavior.

Resolves: #105654
Releases: main, 13.4
Change-Id: Ic12c432c46ba1f3a113e0156fc842f25dbbec1c3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87482
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>