[BUGFIX] Respect aspect ratio for thumbnails in content preview
The thumbnails in the content preview were previously displayed in
a square format, which did not correspond to the aspect ratio of the
image and also led to a distorted display.

The patch corrects the distorted display and ensures that the preview
images are displayed with their correct aspect ratio. The display of
the preview images has also been modernized somewhat by aligning it
with the tile view of the file list.

Resolves: #105845
Releases: main, 13.4
Change-Id: Ic9226300e46f19117fd6858ef6f046aa1d083467
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87497
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>