[BUGFIX] Rework Flexform data handling in TcaFlexPrepare
The previous code in `TcaFlexPrepare->initializeDataValues()` was pretty
optimistic regarding the incoming data and may booted XML parsing on
empty strings and did not consider `null` values.

The code is changed to check now whether the incoming data is not null
and not an empty string before using it at all.

Resolves: #104962
Releases: main, 13.4
Change-Id: If8dbffdc82b1a1e53409e0417d5ddd7a25ed6946
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87684
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>