[BUGFIX] flatpickr: handle proper validation of altInput
flatpickr builds an alternative input field for user interaction if
enabled (which we do) to separate UI date format from system date
formats.
This adds a new layer of complexity in terms of validation: the event
`formengine.dp.change` was dispatched on the original input field, which
is the one being hidden by flatpickr, causing inproper validation of
date picker fields.

The validity state of the validated input field needs to be synchronized
with the `altInput` generated by flarpickr, therefore a custom handler
for the event `t3-formengine-postfieldvalidation` is registered.

Resolves: #105829
Releases: main, 13.4
Change-Id: I456fb8cb60322dde718583ed53a77b33f2f7b09d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87502
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>