[BUGFIX] Fix behavior and styling of EXT:form element type listing
With #104901, the migration from <a> to <button> the JavaScript
handlers were not updated to accommodate the new markup,
preventing proper insertion of elements into form creating with EXT:form.

Additionally, the styling was misaligned for elements and page types
in the form.

This patch restores the correct visual styling for the element list and
updates the JavaScript handlers to ensure compatibility with the new
button-based markup.

Resolves: #105929
Related: #104901
Releases: main, 13.4
Change-Id: I88a1372b721ea6a2833da010453d65b2150e8961
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87752
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>