[BUGFIX] Fix missing FormEngine file upload controls for maxitems=X
Unlike the IRRE inline container, the file upload container uses
a webcomponent, thus the CSS selector with ':scope >' is not
applicable.

Because the selector did not match, the file controls could not
be re-enabled once it was hidden due to a max/minitems configuration.
Also, the form could not be saved because a file would then be
missing.

IRRE childs with max/minitems were not affected.

Resolves: #105938
Related: #105808
Releases: main, 13.4
Change-Id: Ifa798722841a5b0f36345ffd1bede3fc26010979
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87776
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>