[TASK] Refine site set error handling
Show errors for all site set errors in site overview and
site:sets:list console commands.

It turned out that site set errors for site sets that have not
yet been assigned to a site are hard to figure out for integrators by
using the TYPO3 error log, as we only show errors for site sets that
have been assigned in the site overview and settings module.

Therefore all site set errors are streamlined
to be shown in the site overview and command listings and to contain
additional information for the cause of the failure:
 * Invalid or missing properties are reflected in order to understand
   why settings or set definitions fail to be loaded
 * YAML parse errors are now listed
 * Chained dependency errors are now detected and reported

Empty settings.yaml in sets are no longer considered an error
in order for extension templates to be able to ship this file
with comments, but without any concrete contents.

Resolves: #105793
Resolves: #105792
Releases: main, 13.4
Change-Id: Ia18df51ef9ff87217f8d555a3b72692c65a4e1ff
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88058
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>