[BUGFIX] Provide legacy initialization method for the color picker
The color picker is not explicitly marked as internal and was
therefore used by extension for custom backend modules.

With the color picker refactoring in #105552 the JavaScript module
was refactored to use a web component instead of a initializer
method, causing extension to break which import this module.

Provide the legacy interface with an implementation that relies
on the new webcomponent based API.

Resolves: #105963
Related: #105552
Releases: main, 13.4
Change-Id: I14ee616e4ecfa6888fc65a5b253135a5e6a7d1f3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87807
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>