[TASK] Show only file name in file reference header button
Previously, the file name and - if present - the file title
were displayed in file reference's header row.

Both were preceded by a label that indicated whether it
is the file name or title (even if there was no file title).

Depending on the length of the file name and the file title,
this took up different amounts of space. In addition, the
space available in .form-irre-header-button was also influenced
by the ratio of the preview image.

The ratio of the preview image gives the user an important
indication of which format it is, so this is retained. However,
to make the display clearer, now only the file name is displayed
above the thumbnail. This is in line with the display in the
file list module. Additionally, the labels, used to distinguish
between the file name and title are removed from the view.

Resolves: #101949
Releases: main, 13.4
Change-Id: If0764e890e8b2527923fb6bd79ca831948d195a1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87817
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>