[!!!][TASK] Remove TCA interface settings for list view
The TCA options ['interface']['maxSingleDBListItems']
and ['interface']['maxDBListItems'] are removed and not
evaluated anymore.

Thus, the ['interface'] array key is not in use
anymore for TCA, making TCA slimmer, as it is loaded
on every PHP call.

Overriding visual settings can still be done on a per -
User TSconfig or per - Page TSconfig level, which is much
more flexible anyways.

Option "maxSingleDBListItems" is removed in
favor of "mod.web_list.itemsLimitSingleTable"
Option "maxSingleDBListItems" is removed in
favor of "mod.web_list.itemsLimitPerTable"

Resolves: #106412
Releases: main
Change-Id: Ic342b048bbcfdc13ee7cf2e1d9175a96e214d08e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87852
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>