[TASK] Break service loop in ContentFetcher
ContentFetcher is part of the "fluid based page module".
This part of the page module is very hard to stabilize and
to maintain. We had about half a dozen patches to the
"translate button" alone, all boiling down to the fact that
the codebase complexity is so high and so convoluted that
we have a really hard time dealing with it.

Quite a few things went wrong when the surrounding class
architecture has been built. One major issue is that many
classes blend data object and service characteristics
and happily "pollute" service class constructors with
manual data. Most objects assigned to fluid are a
mixture of a service and a data object.

One particularly unfortunate detail is resolved with the
patch: ContentFetcher and PageLayoutContext know each
other - there is an object loop between the two:
ContentFetcher receives PageLayoutContext in __construct(),
parks it in a property, and PageLayoutContext creates an
instance of ContentFetcher handing over itself, and carries
it in a property around as well. ContentFetcher is in fact
a service, it shouldn't have state at all, while
PageLayoutContext is more a data object that shouldn't deal
with services.

To break this loop, the patch makes ContentFetcher readonly
and stateless: The manual constructor argument
PageLayoutContext is removed and hand over to single service
methods instead. This frees __construct() and makes it
usable for proper dependency injection of dependent services.

With the relation between these two key classes being
loosened, we can continue to look at improvements of
PageLayoutContext with additional patches to turn it into
a proper data object, and we can start looking at related
classes to further reduce complexity.

Resolves: #106025
Releases: main
Change-Id: I21bccff392625e57834f6de11a608dc7cbcec991
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87936
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>