[BUGFIX] Correct Workspace Preview
This patch addresses multiple issues with the
workspace preview.

The visibility of the range slider is now restored
by correcting the markup to be used as intended.
The labels were never meant to be placed inside
`.form-range`, as it is only a wrapper for styling
the element. Additionally, the labels are now
properly hidden when the preview is not in slide
mode. To better align with the buttons,
`.form-range-sm` is added to ensure a similar
appearance in size.

The calculation of the element’s height was invalid,
preventing it from being set to 0 or 100%. The
calculation and surrounding CSS were rewritten to
allow a simpler implementation based on percentages
instead of fixed pixel. This enables the view to
better adapt to viewport changes without requiring
recalculations.

The top bar was reorganized for improved behavior
on mobile devices, consuming less space and
providing more room for the preview. To minimize
layout shifts, the view mode switch and action
buttons are now aligned to the end of the top bar.

The CSS was modernized to simplify adjustments and
respect the user's preference for light and dark
mode in the top bar.

Resolves: #106017
Releases: main, 13.4
Change-Id: I0203fd1d96bd74150e53b270c2d2ca34bc408321
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87940
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Riccardo De Contardi <erredeco@gmail.com>