[BUGFIX] Prevent array to string conversion in backend password reset
Submitting an array instead of a string in the backend password
reset form triggers a warning and 503 error due to
`GeneralUtility::validEmail()` function expecting a string.

This change ensures the submitted email address in the
password reset form is always treated as a string. If a
non-string value is detected, it will be safely reset to an
empty string.

Additionally, one unused variable and two superfluous doc
comments have been removed.

Resolves: #106067
Releases: main, 13.4, 12.4
Change-Id: I5cde87570a948e03e3949d1e64d59c03dbf1cfc5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88578
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>