[BUGFIX] Workaround initialization race condition in tree-module-state
The tree module state event handler for module state
updates is registered in a web components connected
callback. That means there is a timeframe between
event handler registration and tree initialiation,
where module updates can not yet be consumed.

These module updates are skipped for now, but should
be collected in a stack later on to be consumed lazy
(also for events that possibly arrive before the
connectedCallback of the tree-module-state mixin).

Resolves: #106128
Related: #105120
Releases: main, 13.4
Change-Id: I87b41d020bdfe61edde00a5bac319b86eec69726
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88130
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>